+++ hedera-mirror-web3/src/main/java/com/hedera/mirror/web3/state/core/MapReadableKVState.java
@@ -48,7 +48,7 @@ public class MapReadableKVState<K, V> extends ReadableKVStateBase<K, V> {
     }
 
     @Override
-    @Deprecated
+    @SuppressWarnings("deprecation")
     public long size() {
         return backingStore.size();
     }
