+++ hedera-mirror-web3/src/main/java/com/hedera/mirror/web3/state/keyvalue/AbstractReadableKVState.java
@@ -20,7 +20,7 @@ public abstract class AbstractReadableKVState<K, V> extends ReadableKVStateBase<
     }
 
     @Override
-    @Deprecated
+    @SuppressWarnings("deprecation")
     public long size() {
         return 0;
     }
